A big update! We can connect lots of Nemisys/Synapse servers in the same time now! I have tested a lot!
